API change: consider projects with no project-tap.i18n as tap-i18n enabled
From now on we consider projects that installed tap-i18n as enabled
regardless of whether or not they set the project-tap.i18n file.
If project-tap.i18n doesn't exist - we use the default options for that
project.

Since projects developers don't need to set project-tap.i18n to enable
tap-i18n we don't support empty project-tap.i18n any longer. We
supported it as a shortcut to enable tap-i18n but now it is redundant.