Do not build/load supported languages with no translation files
When the user specified a language as a supported language
in his project-tap.i18n, we used to create a unified translation
file for that language even if there were no translation files for
that language (even if it was basically an empty JSON).

Later when a client set that language as the client language we
loaded that file even though it is empty.

From now on, no unified translation file will be created for
supported languages that have no translation files and we
won't attempt to load them.