Update currentRule when changing active pluralExtension
When using addRule, currentRule stays unmodified, so the new rule is not active.

Calling setCurrentLng also did not help, because the language does not change,
and currentRule is already defined.

With this patch, when you addRule it checks if you are modifying the current language.
If it's the case, it clears the currentRule.lng  and calls setCurrentLng.

I'm not sure if this is the best approach. Another approach would be to set currentRule to false
and call setCurrentLng(). Or avoid this change and let people delete the currentRule themselves?
But that would mean 3 steps: addRule(..), currentRule = false, setCurrentLng(..).